realease 1.0.0

+++ js/main.js
@@ -3,17 +3,8 @@ $(function () {
   var $drink = $("#drink");
   var $name = $("#name");
 
-  var orderTemplate = ""+
-      "<li>"+
-      "<p><strong>Name:</strong> {{name}}</p>" +
-      "<p><strong>Drink:</strong> {{drink}}</p>" +
-      "<button data-id='{{id}}' class='remove'>X</button>"
-      "</li>";
-
-
-
-
-
+  var orderTemplate = $("#order-template").html();
+      
   function addOrder(order){
     $orders.append(Mustache.render(orderTemplate, order));
   }
@@ -21,7 +12,6 @@ $(function () {
   //
   //get
   //
-
   $.ajax({
     type: "GET",
     url: "http://rest.learncode.academy/api/adrian/orders",
@@ -39,18 +29,16 @@ $(function () {
   //
   // add order submit
   //
-
   $("#add-order").on("click", function(){
     var order = {
       name: $name.val(),
       drink: $drink.val(),
     };
 
-//
-// post
-//
-
-    $.ajax({
+  //
+  // post
+  //
+  $.ajax({
       type: "POST",
       url: "http://rest.learncode.academy/api/adrian/orders",
       data: order,
@@ -64,24 +52,83 @@ $(function () {
       error: function(){
         alert("error loading orders");
       }
-    });
+  });
 
 
 
   });
 
+  //
+  //delete
+  //
   $orders.delegate('.remove','click', function(){
 
     var $li = $(this).closest('li');
-    $.ajax({
-      type: 'DELETE',
-      url: 'http://rest.learncode.academy/api/adrian/orders/'+ $(this).attr('data-id'),
-      success: function(){
-        $li.fadeOut(300, function(){
-          $(this).remove();
-        });
-      }
-    });
+
+    if (confirm("¿Seguro que deseas eliminar la comanda?")){
+      $.ajax({
+        type: 'DELETE',
+        url: 'http://rest.learncode.academy/api/adrian/orders/'+ $(this).attr('data-id'),
+        success: function(){
+          $li.fadeOut(300, function(){
+            $(this).remove();
+          });
+        }
+      });
+    }
+
   });
+
+  //
+  //edit
+  //
+  $orders.delegate('.editOrder','click', function(){
+
+    var $li = $(this).closest('li');
+    $li.find('input.name').val( $li.find('span.name').html() );
+    $li.find('input.drink').val( $li.find('span.drink').html() );
+    $li.addClass('edit');
+    
+  });
+  
+  //
+  //cancel edit
+  //
+  $orders.delegate('.cancelEdit','click', function(){
+
+    var $li = $(this).closest('li').removeClass('edit');
+
+  });
+
+  //
+  //save edit
+  //
+  $orders.delegate('.saveEdit','click', function(){
+
+    var $li = $(this).closest('li');
+    var order ={
+      name: $li.find('input.name').val(),
+      drink: $li.find('input.drink').val()
+    };
+
+     $.ajax({
+      type: 'PUT',
+      url: 'http://rest.learncode.academy/api/adrian/orders/'+ $li.attr('data-id'),
+      data: order,
+      success : function(newOrder) {
+        $li.find('span.name').html(order.name);
+        $li.find('span.drink').html(order.drink);
+        $li.removeClass('edit');
+        window.location.reload(true);
+      },
+      error: function(){
+        alert("error updating orders");
+      }
+
+    });
+
+  });
+
+
 
 });
